Moved Bonos, Categorias and Pepe to their own files
